Move _create_all_objects into runner.py
