+++ modal/runner.py
@@ -3,7 +3,7 @@ import asyncio
 import dataclasses
 import os
 from multiprocessing.synchronize import Event
-from typing import TYPE_CHECKING, AsyncGenerator, List, Optional, TypeVar
+from typing import TYPE_CHECKING, AsyncGenerator, Dict, List, Optional, TypeVar
 
 from grpclib import GRPCError, Status
 from rich.console import Console
@@ -13,6 +13,7 @@ from modal_proto import api_pb2
 
 from ._output import OutputManager, get_app_logs_loop, step_completed, step_progress
 from ._pty import get_pty_info
+from ._resolver import Resolver
 from ._sandbox_shell import connect_to_sandbox
 from ._utils.app_utils import is_valid_app_name
 from ._utils.async_utils import TaskContext, synchronize_api
@@ -20,7 +21,8 @@ from ._utils.grpc_utils import retry_transient_errors
 from .app import _LocalApp, is_local
 from .client import HEARTBEAT_INTERVAL, HEARTBEAT_TIMEOUT, _Client
 from .config import config, logger
-from .exception import InteractiveTimeoutError, InvalidError, _CliUserExecutionError
+from .exception import ExecutionError, InteractiveTimeoutError, InvalidError, _CliUserExecutionError
+from .object import _Object
 
 if TYPE_CHECKING:
     from .stub import _Stub
@@ -60,9 +62,8 @@ async def _init_local_app_new(
     app_resp = await retry_transient_errors(client.stub.AppCreate, app_req)
     app_page_url = app_resp.app_logs_url
     logger.debug(f"Created new app with id {app_resp.app_id}")
-    return _LocalApp(
-        client, app_resp.app_id, app_page_url, environment_name=environment_name, interactive=interactive
-    )
+    return _LocalApp(client, app_resp.app_id, app_page_url, environment_name=environment_name, interactive=interactive)
+
 
 async def _init_local_app_from_name(
     client: _Client,
@@ -86,6 +87,70 @@ async def _init_local_app_from_name(
         return await _init_local_app_new(
             client, name, api_pb2.APP_STATE_INITIALIZING, environment_name=environment_name
         )
+
+
+async def _create_all_objects(
+    app: _LocalApp,
+    indexed_objects: Dict[str, _Object],
+    new_app_state: int,
+    environment_name: str,
+    output_mgr: Optional[OutputManager] = None,
+):  # api_pb2.AppState.V
+    """Create objects that have been defined but not created on the server."""
+    if not app.client.authenticated:
+        raise ExecutionError("Objects cannot be created with an unauthenticated client")
+
+    resolver = Resolver(
+        app.client,
+        output_mgr=output_mgr,
+        environment_name=environment_name,
+        app_id=app.app_id,
+    )
+    with resolver.display():
+        # Get current objects, and reset all objects
+        tag_to_object_id = app.tag_to_object_id
+        app.tag_to_object_id = {}
+
+        # Assign all objects
+        for tag, obj in indexed_objects.items():
+            # Reset object_id in case the app runs twice
+            # TODO(erikbern): clean up the interface
+            obj._unhydrate()
+
+        # Preload all functions to make sure they have ids assigned before they are loaded.
+        # This is important to make sure any enclosed function handle references in serialized
+        # functions have ids assigned to them when the function is serialized.
+        # Note: when handles/objs are merged, all objects will need to get ids pre-assigned
+        # like this in order to be referrable within serialized functions
+        for tag, obj in indexed_objects.items():
+            existing_object_id = tag_to_object_id.get(tag)
+            # Note: preload only currently implemented for Functions, returns None otherwise
+            # this is to ensure that directly referenced functions from the global scope has
+            # ids associated with them when they are serialized into other functions
+            await resolver.preload(obj, existing_object_id)
+            if obj.object_id is not None:
+                tag_to_object_id[tag] = obj.object_id
+
+        for tag, obj in indexed_objects.items():
+            existing_object_id = tag_to_object_id.get(tag)
+            await resolver.load(obj, existing_object_id)
+            app.tag_to_object_id[tag] = obj.object_id
+
+    # Create the app (and send a list of all tagged obs)
+    # TODO(erikbern): we should delete objects from a previous version that are no longer needed
+    # We just delete them from the app, but the actual objects will stay around
+    indexed_object_ids = app.tag_to_object_id
+    assert indexed_object_ids == app.tag_to_object_id
+    all_objects = resolver.objects()
+
+    unindexed_object_ids = list(set(obj.object_id for obj in all_objects) - set(app.tag_to_object_id.values()))
+    req_set = api_pb2.AppSetObjectsRequest(
+        app_id=app.app_id,
+        indexed_object_ids=indexed_object_ids,
+        unindexed_object_ids=unindexed_object_ids,
+        new_app_state=new_app_state,  # type: ignore
+    )
+    await retry_transient_errors(app.client.stub.AppSetObjects, req_set)
 
 
 @asynccontextmanager
@@ -156,7 +221,7 @@ async def _run_stub(
         exc_info: Optional[BaseException] = None
         try:
             # Create all members
-            await app._create_all_objects(stub._indexed_objects, app_state, environment_name, output_mgr=output_mgr)
+            await _create_all_objects(app, stub._indexed_objects, app_state, environment_name, output_mgr=output_mgr)
 
             # Update all functions client-side to have the output mgr
             for obj in stub.registered_functions.values():
@@ -191,7 +256,9 @@ async def _run_stub(
                     logs_loop.cancel()
             else:
                 output_mgr.print_if_visible(
-                    step_completed(f"App aborted. [grey70]View run at [underline]{app.app_page_url}[/underline][/grey70]")
+                    step_completed(
+                        f"App aborted. [grey70]View run at [underline]{app.app_page_url}[/underline][/grey70]"
+                    )
                 )
                 output_mgr.print_if_visible(
                     "Disconnecting from Modal - This will terminate your Modal app in a few seconds.\n"
@@ -236,8 +303,8 @@ async def _serve_update(
 
         # Create objects
         output_mgr = OutputManager(None, True)
-        await app._create_all_objects(
-            stub._indexed_objects, api_pb2.APP_STATE_UNSPECIFIED, environment_name, output_mgr=output_mgr
+        await _create_all_objects(
+            app, stub._indexed_objects, api_pb2.APP_STATE_UNSPECIFIED, environment_name, output_mgr=output_mgr
         )
 
         # Communicate to the parent process
@@ -321,8 +388,8 @@ async def _deploy_stub(
 
         try:
             # Create all members
-            await app._create_all_objects(
-                stub._indexed_objects, post_init_state, environment_name=environment_name, output_mgr=output_mgr
+            await _create_all_objects(
+                app, stub._indexed_objects, post_init_state, environment_name=environment_name, output_mgr=output_mgr
             )
 
             # Deploy app
@@ -332,7 +399,9 @@ async def _deploy_stub(
                 name=name,
                 namespace=namespace,
                 object_entity="ap",
-                visibility=(api_pb2.APP_DEPLOY_VISIBILITY_PUBLIC if public else api_pb2.APP_DEPLOY_VISIBILITY_WORKSPACE),
+                visibility=(
+                    api_pb2.APP_DEPLOY_VISIBILITY_PUBLIC if public else api_pb2.APP_DEPLOY_VISIBILITY_WORKSPACE
+                ),
             )
             try:
                 deploy_response = await retry_transient_errors(client.stub.AppDeploy, deploy_req)
